hopefully fixing fluency_timed bug

+++ fluency/models.py
@@ -85,7 +85,7 @@ class Task(models.Model):
  
 class FluencyUntimedScore(models.Model):
     id              = models.AutoField(primary_key=True, editable=False)
-    score           = models.DecimalField(blank=True, null=True, decimal_places=3,max_digits=4)
+    score           = models.DecimalField(blank=True, null=True, decimal_places=3,max_digits=5)
     unit_id         = models.ForeignKey(Unit, on_delete=models.CASCADE)
     deck_id         = models.ForeignKey(Deck, on_delete=models.CASCADE)
     date            = models.DateTimeField()
@@ -93,8 +93,8 @@ class FluencyUntimedScore(models.Model):
 
 class FluencyTimedScore(models.Model):
     id              = models.AutoField(primary_key=True, editable=False)
-    score           = models.DecimalField(blank=True, null=True, decimal_places=3,max_digits=4)
-    time_elapsed    = models.DecimalField(blank=True, null=True, decimal_places=3,max_digits=4) # time elapsed to complete quiz in seconds
+    score           = models.DecimalField(blank=True, null=True, decimal_places=3,max_digits=5)
+    time_elapsed    = models.IntegerField(blank=True, null=True) # time elapsed to complete quiz in seconds
     unit_id         = models.ForeignKey(Unit, on_delete=models.CASCADE)
     deck_id         = models.ForeignKey(Deck, on_delete=models.CASCADE)
     date            = models.DateTimeField()
